+++ packages/moonstone/ContextualPopupDecorator/ContextualPopupDecorator.js
@@ -113,6 +113,11 @@ const ContextualPopupDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			}
 		}
 
+		componentWillUnmount () {
+			this.containerNode = null;
+			this.clientNode = null;
+		}
+
 		getContainerPosition (containerNode, clientNode) {
 			let position = {};
 			switch (this.adjustedDirection) {